+++ src/tetromino.hpp
@@ -32,13 +32,13 @@ inline Rotation operator+(Rotation rotation, const int offset) {
     }
 
     if (offset > 0) {
-        for (int i = 0; i < offset; ++i) {
+        for (usize i = 0; i < static_cast<usize>(offset); ++i) {
             ++rotation;
         }
         return rotation;
     }
 
-    for (int i = 0; i < -offset; ++i) {
+    for (usize i = 0; i < static_cast<usize>(-offset); ++i) {
         --rotation;
     }
     return rotation;
